\name{serve}
\alias{serve}
\alias{callback}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Starting the triggr server}
\description{...}
\usage{
...
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{callback}{...}
  \item{port}{...}  
}

\details{...}
\note{...}
\value{...}

\author{
Miron B. Kursa <M.Kursa@icm.edu.pl>
}


