\name{triggr-package}
\alias{triggr-package}
\alias{triggr}
\docType{package}
\title{Triggr}

\description{

}
\details{
\tabular{ll}{
Package: \tab triggr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-05-21\cr
License: \tab LICENCE\cr
LazyLoad: \tab yes\cr
}

Triggr is...


}
\author{
Miron B. Kursa <M.Kursa@icm.edu.pl>
}

\keyword{package}
